+++ Sources/KeyboardKit/_Deprecated/KeyboardInputSetProvider.swift
@@ -1,0 +1,19 @@
+import Foundation
+
+@available(*, deprecated, renamed: "InputSetProvider")
+public typealias KeyboardInputSetProvider = InputSetProvider
+
+@available(*, deprecated, renamed: "EnglishInputSetProvider")
+public typealias EnglishKeyboardInputSetProvider = EnglishInputSetProvider
+
+@available(*, deprecated, renamed: "LocalizedInputSetProvider")
+public typealias LocalizedKeyboardInputSetProvider = LocalizedInputSetProvider
+
+@available(*, deprecated, renamed: "StandardInputSetProvider")
+public typealias StandardKeyboardInputSetProvider = StandardInputSetProvider
+
+@available(*, deprecated, renamed: "StaticInputSetProvider")
+public typealias StaticKeyboardInputSetProvider = StaticInputSetProvider
+
+@available(*, deprecated, renamed: "PreviewInputSetProvider")
+public typealias PreviewKeyboardInputSetProvider = PreviewInputSetProvider